Web MIDI: stop sending messages when the document is going to be closed
When the page is going to be closed, loader() may return nullptr.
In MIDIAccess::sendMIDIData(), we refer the DocumentLoader without any
check, thus it results in a renderer crash though it rarely happens.

I had a report on a similar issue, but the specific code in the case was
already removed. This patch also fixes another potential problem that
may cause the same issue.

BUG=674005

Review-Url: https://codereview.chromium.org/2566223006
Cr-Original-Commit-Position: refs/heads/master@{#438485}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be3f5b8ec3199a8c6fbba72cbc83776d6ae53efc