Rename OnError to OnErrorAdapter and OnSuccess to OnSucessAdapter.
The rename is needed to avoid a naming collision after changing from
Blink to Chromium naming style.  Right now we have an |OnError| type and
an |onError| virtual method (differing by case of the first character);
after a naive rename by the rewrite_to_chrome_style tool we would end up
with |OnError| being the name of both the type and the method (with both
living in the same namespace). Similar situation happens for onSuccess /
OnSuccess names.

Changing the type name is a workaround that fits into the guidance on
the recommended post-Blink-to-Chromium-rename style suggested by
esprehn@ in https://crbug.com/582312#c17:
- Getters favor not using "Get", ex. FirstChild()
- Unless the type name conflicts, in which case you can either rename
  the type if it's easy and makes sense, or add "Get", ex. GetContext().

BUG=673480

Review-Url: https://codereview.chromium.org/2570193002
Cr-Original-Commit-Position: refs/heads/master@{#438579}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 18c69c35b756dca9096cf2fc0ab69d4b6d7ddba5