Tighten expectations for 3 flaky animated compositing tests.
These 3 tests are flaky when run in random order or even alone.
As a first step in improving them, modify the tests to get away
from setTimeout and to verify that the pauseAnimationAPI is working.

This is unlikely to resolve the flakiness, but it will help narrow
the cause.

R=chrishtr
BUG=653722

Review-Url: https://codereview.chromium.org/2569063002
Cr-Original-Commit-Position: refs/heads/master@{#438608}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2e4858e7eef2135034ed003e58a380e0813da9e6