Moved action verbs out of Blink.
When Blink is asked to provide what actions an accessibility object supports, instead of returning a string with an action verb, it now returns a value from the |WebAXSupportedAction| enumeration.
This change allows different platforms to localize the actions in any way they please. More specifically, on Windows we now expose two APIs: one for returning the localized action verb and one that returns the verb in English. This is a requirement dictated by the IAccessible2 Spec.
BUG=640043
TESTED=By changing Chrome's language to something other than English and trying to read a clickable element with NVDA
R=dmazzoni@chromium.org, sky@chromium.org, avi@chromium.org

Review-Url: https://codereview.chromium.org/2518183002
Cr-Original-Commit-Position: refs/heads/master@{#438649}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6e043304856897ee42289b94181baf34726f83c2