Use random URLs in tracing layout tests to avoid caching
Some layout tests in http/tests/inspector/tracing check |encoded_data_length|
which is not dispatched when the resource is cached because no bytes are
received from the network in such a case. To deflake them, this CL adds a
random query to avoid caching.

BUG=None
R=dgozman@chromium.org

Review-Url: https://codereview.chromium.org/2573393002
Cr-Original-Commit-Position: refs/heads/master@{#438752}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 86d86afecbe179c6da65db526397216ce6cd6f1d