Remove performanceMonitor()->unsubscribeAll() from Performance's destructor
Performance object is on Oilpan's heap, so its destructor is not allowed to touch
other on-heap objects.

Also we don't need to call performanceMonitor()->unsubscribeAll() because it is
just unregistering a back pointer Member to the Performance object from the map
in PerformanceMonitor.

BUG=

Review-Url: https://codereview.chromium.org/2584733003
Cr-Original-Commit-Position: refs/heads/master@{#439081}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 18502d9c2a5ec6c54e8f7cd2a11bb2e45e78e331