Dispose a persisted plugin widget if the content frame is detached.
HTMLPlugInElement::removedFrom() clears persisted plugin widgets, to
prevent reuse of a widget associated with a destroyed object. However,
when nodes containing connected subframes are removed from a Document,
it is possible to synchronously trigger script after subframes are
detached, but before the nodes themselves are actually removed. This
can cause a FrameView for a detached frame to be incorrectly reused.

BUG=671331

Review-Url: https://codereview.chromium.org/2586633002
Cr-Original-Commit-Position: refs/heads/master@{#439339}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9485e0d3ee1ec3d8bd5578b7ed4493855cfc9898