Specify TaskType of posted Task explicitly in Document
This CL add explicit TaskTypes for tasks posted by Document itself.
On Document::addConsoleMessage, since it's an Inspector message,
this should not be throttled or suspended, so its TaskType should be
Unthrottled.
On Document::setAutofocusElement, the TaskType should be UserInteraction
according to the spec:
https://html.spec.whatwg.org/multipage/forms.html#autofocusing-a-form-control:-the-autofocus-attribute

BUG=624696

Review-Url: https://codereview.chromium.org/2583003002
Cr-Original-Commit-Position: refs/heads/master@{#439381}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eda496bacb9a4b17c3995473a3725bb5d19559f0