Install supplements at every navigation
Currently supplement objects are reused when an initial empty document navigates to
a same-origin document. This is problematic because ContextLifecycleObservers are
detached when the initial empty document is destroyed. They cannot be used after
the navigation.

To fix the problem, we need to install supplements at every navigation.

BUG=610176

Review-Url: https://codereview.chromium.org/2580753004
Cr-Original-Commit-Position: refs/heads/master@{#439413}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dd2e0942b294948ee891b1e54ca17a195fc1f3a1