Plumb site engagement to the renderer process.
This CL generalises the EngagementLevel concept in site engagement,
moving it to a Mojo enum in Blink's public/platform. Prior to each
navigation commit, and when the engagement level for an origin changes
between buckets, the browser process sends the new level to all renderers
which are navigated to that origin. Using the discrete levels rather
than the continuous score minimises the IPC overhead to one additional
IPC per navigation commit per frame, and one IPC per engagement level
change.

The engagement level is stored on Document within the renderer, with
a temporary stop on RenderFrameImpl if sent prior to navigation commit.
This is a first step to enabling the collection of UseCounter data for
JavaScript API usage against engagement buckets.

BUG=655415

Review-Url: https://codereview.chromium.org/2535483002
Cr-Original-Commit-Position: refs/heads/master@{#439726}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6c1f1cf3c31b4c798708cb17d4ecd8e6db37e833