Better handle non-existant wrappers when intercepting writes to plugin objects
Creating a wrapper might fire events synchronously which in turn can
modify the wrapper. There is, however, an invariant for interceptors
that if they modify their holder, they need to signal this. Therefore,
when we create a wrapper, always signal that the write happened.

BUG=670155
R=tkent@chromium.org,haraken@chromium.org

Review-Url: https://codereview.chromium.org/2566663002
Cr-Original-Commit-Position: refs/heads/master@{#439812}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 806a7d2250a82acc584e83cf9086701f0e77cf6d