DevTools: teach TabbedPane.setTabIcon to accept UI.Icon instances
This patch is a preparation for a "badged icons" instead of green
checkmarks in the "persistence2" experiment.

Currently, TabbedPane.setTabIcon method accepts icon type and icon
titles. This works fine, but we'd like to pass in a more complex
objects - badged icons.

For this to work, the TabbedPane.setTabIcon has to accept general
instances of UI.Icon class.

BUG=675856
R=dgozman

Review-Url: https://codereview.chromium.org/2587293002
Cr-Original-Commit-Position: refs/heads/master@{#439828}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 911cf4b74f6dbce780c83e31b51358bb2c849416