third_party: Add libdrm.
libdrm shipped with the sysroot or our workstation is older than libdrm
we use on chromeos.
This forces us to redefine symbols/defines (fourcc code for example) and
doesn't allow us to compile newer drm code on trybots (drm atomic).

This CLs adds libdrm to third_party, so that we can build the third_party
one when we build on our workstation or on the trybot, similar to what
we currently do with minigbm.

BUG=

Review-Url: https://codereview.chromium.org/2594483002
Cr-Original-Commit-Position: refs/heads/master@{#439907}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 141830582db701ab4edf529ae876906a6508cf7a