PaymentApp: Remove scope_url parameter from Get/SetManifest methods.
When calling setManifest()/getManifest() from Blink side, we currently pass
scope_url to find associated service worker in Content layer. However, it is
not a good way to pass scope_url each time. Of course, since the implementations
are all private methods, there is no risk of calling it wrong from the outside.
But some test classes are declared as friend class and it can be misused in the
classes.

Therefore, this CL is adding Init() in payment_app.mojom and then using it in
Blink side PaymentAppManager when creating the object. Then, we don't need to
pass scope_url each time when calling setManifest()/getManifest().

BUG=661608
TEST=existing tests

Review-Url: https://codereview.chromium.org/2586203003
Cr-Original-Commit-Position: refs/heads/master@{#439942}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f87f3d68d3a82070b599948cf6fbed3b224e0bf0