Rename CheckIntersectionOrEnclosure enum type to GeometryMatchingMode.
This CL renames blink::SVGSVGElement::CheckIntersectionOrEnclosure enum
type to GeometryMatchingMode.

The rename is needed to avoid a naming collision after changing from
Blink to Chromium naming style.  Right now we have an
|CheckIntersectionOrEnclosure| enum type and a
|checkIntersectionOrEnclosure| method (differing by case of the first
character); after a naive rename by the rewrite_to_chrome_style tool we
would end up with |CheckIntersectionOrEnclosure| being the name of both
the type and the method (with both living in the same namespace).

Changing the type name is a workaround that fits into the guidance on
the recommended post-Blink-to-Chromium-rename style suggested by
esprehn@ in https://crbug.com/582312#c17:
- Getters favor not using "Get", ex. FirstChild()
- Unless the type name conflicts, in which case you can either rename
  the type if it's easy and makes sense, or add "Get", ex. GetContext().

BUG=582312

Review-Url: https://codereview.chromium.org/2596503002
Cr-Original-Commit-Position: refs/heads/master@{#439947}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 89d3dbf3be30f5a4b98734e5c4089c874d97c530