Changed EVerticalAlign to an enum class and renamed its members
Changed EVerticalAlign to an enum class and gave it an unsigned
underlying type.

Changing it to an enum class enforces better namespacing and code
practices. Adding the unsigned underlying type is pre-work for when the
class is eventually stored as an enum bitfield (it would be done in this
patch, except a presubmit warning already exists that prevents that. The
presubmit warning needs to be updated before that change can occur.)

This is also pre-work to move EVerticalAlign to be generated in
ComputedStyleBase.

BUG=628043

Review-Url: https://codereview.chromium.org/2578403002
Cr-Original-Commit-Position: refs/heads/master@{#440009}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2f0c2019085dfcf95005bb54b9eb483893e8cb1d