Improve separation between "SMIL times" and timestamps
The 'elapsed' time does not need to be a SMILTime in general - it can
only be non-finite in the case where the document is not active, and
then the timeline should not be running/animations updated.
Thread the double value further from SMILTimeContainer down into
SVGSMILElement. Simplify some computations.

BUG=641437,676137

Review-Url: https://codereview.chromium.org/2592103002
Cr-Original-Commit-Position: refs/heads/master@{#440171}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d6b6ebf86773fcafd3ce6cdc6fe7ef7d14bae91e